libcap-ng: add livecheck, update homepage
By default, livecheck isn't able to check the `stable` URL for
`libcap-ng`, so it falls back to checking the Git tags from the `head`
URL. This adds a `livecheck` block that checks the homepage, which
links to the `stable` tarball. This also updates the `homepage` URL
to add a trailing forward slash, to avoid an unnecessary redirection.