libcdio-paranoia 10.2-2.0.2 (new formula)
This is a cross-platform version of cdparanoia using libcdio for disc
access. Unlike cdparanoia (now removed from Homebrew), it is maintained.

Testing is challenging since functionality requires the disc drive. Per
advice in https://docs.brew.sh/Formula-Cookbook#add-a-test-to-the-formula,
this simply tests that the error is correct.

This is a redo of #27347 and #114860.