zbar: update livecheck
The existing `livecheck` block for `zbar` was added at a time when the
formula used a GitHub tag tarball but it's currently using a tarball
from linuxtv.org. The upstream repository also contains a link to
https://linuxtv.org/downloads/zbar/, so this updates the `livecheck`
block to check that directory listing page to align the check with
the `stable` source.