tailwindcss: add livecheck comment and head url
The `livecheck` block for `tailwindcss` uses the `GithubLatest`
strategy but the `stable` URL is a tarball from npm. This adds a
comment before the `livecheck` block to explain the exception.

This also adds a `head` URL and replaces the `livecheck` block URL
with `:head`. `brew install --HEAD` worked fine when I tested it
locally.