qsoas: update livecheck
The `livecheck` block for `qsoas` originally checked the `stable`
tarball link on the first-party download page but it was later
modified to check the Git repository tags without any explanation for
why. The download page links to an up to date tarball, so it doesn't
seem like checking the Git tags is necessary.

This updates the `livecheck` block to check the first-party download
page again, which aligns the check with the `stable` source.