digdag: update stable, livecheck
The existing `livecheck` block for `digdag` checks the GitHub
repository's tags, presumably because the first-party website doesn't
make it easy to identify the newest version. The homepage directs
users to https://dl.digdag.io/digdag-latest, which redirects to the
jar file from the latest GitHub release (it doesn't redirect to the
versioned dl.digdag.io URL we use as `stable`). The release notes page
in the documentation (https://docs.digdag.io/releases.html) doesn't
include versions with four parts like 0.10.5.1 (only 0.10.5).

The formula's `stable` URL uses dl.digdag.io URL, which simply
redirects to the jar file from the corresponding GitHub release.
Outside of the `digdag-latest` URL, the first-party website doesn't
appear to link to any other dl.digdag.io URLs.

With that in mind, this updates the `stable` URL to use the jar file
from the latest release asset (the `sha256` is unchanged because it's
the same file from the same source) and modifies the `livecheck`
block to use the `GithubLatest` strategy (since we're using a GitHub
release asset).