crash: update livecheck
The existing `livecheck` block for `crash` is an old check from years
ago and there's no specific reason it's checking Git tags other than
it technically works. This updates the `livecheck` block to use the
typical approach we use for maven.org formulae (I'll create a
strategy for this in due time).