di: update livecheck
The `livecheck` block for `di` was recently modified to check loose
version text on the homepage, as the existing `Sourceforge` check was
returning an unstable 4.99.5 version from the `/beta` subdirectory as
newest. This addresses the issue by updating the previous
`Sourceforge` check's regex to only target matching tarballs in the
root directory.