basex: update livecheck
The existing `livecheck` block for `basex` works but the first-party
download page links to the file we use in the `stable` URL, so it may
be better to check that. Theoretically, livecheck may surface a new
version before the download page is updated if upstream uploads the
files first, so checking the download page may be a more reliable
indicator that a version is released (if it's kept up to date).