flagd: add livecheck
We noticed in a recent version bump PR for `flagd` that upstream
only appears to mark `core/v1.2.3` releases as "latest", not releases
like `flagd/v1.2.3`. This adds a `livecheck` block that checks the
`flagd/v1.2.3` tags, omitting unrelated tags. This also adds a comment
to explain the situation, so hopefully maintainers will know not to
wait for a `flagd` version to be marked as "latest".