squid 6.13
This updates `squid` to the latest stable version, 6.13.

The `squid` website now links to the GitHub repository as the source
for new versions. Upstream did not publish a `squid-6.13.tar.xz` file
on the first-party website (the URL simply redirects to the
`/Versions/` page), so I've switched to the bootstrapped tarball
that's published as a GitHub release asset for 6.13.

This also updates the `livecheck` block to use the `GithubLatest`
strategy, as there are Git tags for a higher major version that use
the same tag format as the current release series (e.g.,
`SQUID_7_0_1`).