ltl2ba: add livecheck
By default, livecheck is unable to identify version information for
`ltl2ba`. The first-party download page only provides a link to the
latest tarball after submitting a form, so we haven't been able to
add a check for this because it requires a `POST` request. I recently
added `POST` support to livecheck and it is now in a tagged `brew`
release, so we can use it in `livecheck` blocks.

This adds a `livecheck` block that makes a `POST` request, using the
minimum form values required to receive the download page content.