rakudo-star: add livecheck
By default, livecheck checks the Git tags for `rakudo-star` but it's
returning 2025.02 as newest instead of 2025.01. The 2025.02 tag was
created six days ago but a related GitHub release hasn't been created
yet. This adds a `livecheck` block using the `GithubLatest` strategy,
as the `stable` URL is a GitHub release asset.