sdl2_ttf: update livecheck
The `sdl2` formula's `livecheck` block was updated to check the Git
tags instead of GitHub releases, as there are now SDL3 versions but
we need to specifically match SDL2 versions. The `stable` URL is a
GitHub release asset and using the `GithubReleases` strategy would
work for now but eventually it may fail if/when there aren't any SDL2
versions in the most recent releases, so we're using the Git strategy
instead. This updates the `livecheck` block for `sdl2_ttf` to bring
it in line with `sdl2`.