selenium-server 4.29.0
The jar file for `selenium-server` 4.28.1 was published as a GitHub
release asset on the 4.28.0 release and the autobump workflow isn't
able to update to 4.29.0 because the URL substitution results in a
broken URL (i.e., it replaces 4.28.1 in the filename but not 4.28.0
in the tag). This manually updates the formula to 4.29.0.