workflows/publish-commit-bottles: handle PRs that only remove formulae
Previously, we checked only if a PR does not touch any formulae before
we decide that it has no bottles to publish.

After this change, we now check that a PR touches a formula and does not
remove at least one of them before deciding that a PR does not have any
bottles to publish.

This fixes failed bottle publish attempts like what is seen at #209657.