core-lightning: update livecheck
The existing `livecheck` block for `core-lightning` uses the Git
and this is currently returning 25.02 as the newest version but there
is no corresponding release after the better part of a day. The
`stable` URL is a GitHub release asset, so this updates the
`livecheck` block to use the `GithubLatest` strategy, which correctly
returns 24.11.1 as the latest version.