evince 48.0
`evince` is in the autobump list but the URL isn't able to be
automatically updated for a new major version because the URL also
contains the major version in the path (i.e., we end up with a URL
that has the correct filename but uses the previous major version in
the path, so it predictably fails). This manually updates the formula
to 48.0, to get it back on track.