gcr 4.4.0.1
`gcr` is in the autobump list but tarballs are kept in version
directories that only use the major/minor and that part of the path
isn't automatically modified when updating the URL, so autobump fails
for a new major/minor version. This manually updates the formula to
the newest version.