jbang 0.125.0
By default, livecheck checks the Git tags for `jbang` and it's
giving 0.125.1 as the newest version but the formula uses a GitHub
release asset and there isn't a corresponding release after eight
hours. This updates the formula to the newest release, 0.125.0, and
adds a `livecheck` block using the `GithubLatest` strategy, as there
can be a notable gap between tag and release.

Signed-off-by: Rui Chen <rui@chenrui.dev>

Co-authored-by: Sam Ford <1584702+samford@users.noreply.github.com>