ruff-lsp: add livecheck
The `ruff-lsp` formula is deprecated (as the package is deprecated
upstream) but it seems that upstream may still be publishing new
versions. This adds a minimal `livecheck` block to allow us to
continue checking for new versions despite the deprecation, until it's
clear that there won't be further versions. At that point, we can
remove the `livecheck` block and remove `ruff-lsp` from the autobump
list.