qsoas: remove from autobump list
`qsoas` is in the autobump list but the `livecheck` block has been
consistently failing in the autobump environment (Ubuntu 22.04). The
check works fine on macOS but it produces a curl error on Linux
(`curl: (60) SSL certificate problem: unable to get local issuer
certificate`), as the server has an incomplete certificate chain.

This removes `qsoas` from the autobump list and adds an explanatory
comment before the `livecheck` block (so we won't try to autobump the
formula until this is resolved upstream). This check will fail on Linux
CI but technical issues stand in the way of only having the
`livecheck` block in an `on_macos` block, so we will have to skip
livecheck on those PRs for now. [Namely, `brew style` will complain
if we wrap the `livecheck` block in an `on_macos` block
(`FormulaAudit/ComponentsOrder: Nest on_macos blocks inside livecheck
blocks when there is only one inner block`) but if we do that we get
an `undefined method 'on_macos' for an instance of Livecheck` runtime
error.]