lha: remove due to non-free license
The license was incomplete for this formula. As mentioned by the
GitHub repo description:

> Note: This software is licensed under the ORIGINAL LICENSE.
> It is written in man/lha.n in Japanese(Translated:
> <https://packages.debian.org/changelogs/pool/non-free/l/lha/current/copyright>)

As seen in above url, Debian family of distros put `lha` under non-free
which means it does not meet the homebrew/core license requirement[^1].

Other repositories have also applied similar analysis of license:
* Nixpkgs have considers it "unfree" due to old code under LHa license
  having additional requirements on commercial use[^2].
* MacPorts considers license Restrictive/Distributable[^3] which is
  often what distributions classify as non-free.
* Fedora does not provide `lha` but the RPM Fusion includes it inside
  nonfree/release with LicenseRef-Fedora-UltraPermissive as part of
  license[^4].

`lha` can be reconsidered if the current maintainer is able to remove
all old code under the prior license. Users may want to consider `lhasa`
which was created as a free alternative.

[^1]: https://docs.brew.sh/License-Guidelines
[^2]: https://github.com/NixOS/nixpkgs/blob/master/pkgs/by-name/lh/lha/package.nix#L29-L34
[^3]: https://github.com/macports/macports-ports/blob/master/archivers/lha/Portfile#L18
[^4]: https://download1.rpmfusion.org/nonfree/fedora/development/rawhide/Everything/source/SRPMS//repoview/lha.html