orientdb: update homepage, livecheck
The existing `homepage` and `livecheck` block URLs for `orientdb`
redirect to the GitHub organization. In the case of the `livecheck`
block, this causes the check to fail.

This updates the `homepage` to the GitHub project and modifies the
`livecheck` block to use the `GithubLatest` strategy, matching the
version from the related Maven link in the release description.