f2: downgrade to 2.0.3, add livecheck
An upstream v2.1.0 tag for `f2` existed yesterday and autobump picked
it up but it has since been removed. I'm not sure if there was a 2.1.0
release on GitHub at the time but there isn't one now. Either upstream
created the v2.1.0 tag and removed it without creating a release or
they yanked the version.

Either way, 2.1.0 doesn't exist right now, so this downgrades the
formula to 2.0.3. This also adds a `livecheck` block that uses the
`GithubLatest` strategy, so hopefully we will avoid this situation in
the future.