gnutls: update livecheck
The existing `livecheck` block for `gnutls` checks the first-party
news page but this hasn't been updated in a year, so it doesn't
contain information for versions 3.8.5 to 3.8.9. The first-party
download page links to directory listing pages for the "Next" and
"Current stable" releases, so we have to check the download page to
identify the newest major/minor version and then check the directory
listing page to identify the full version from the tarball URLs.