ssocr: add livecheck
By default, livecheck checks the Git tags for `ssocr` and this works
but the `stable` tarball isn't from GitHub. Instead of hoping that
the related tarball exists whenever a Git tag is created, we can
simply match the version from the tarball link on the homepage. This
adds a `livecheck` block to do so, aligning the check with the
`stable` source.