unoconv: deprecate
The upstream GitHub repository for `unoconv` was archived on
2025-03-31. The `README` was updated on 2021-11-12 with a message
saying that Unoconv is deprecated in favor of Unoserver (a rewrite)
and that Unoconv will become unsupported once Unoserver has the major
features of Unoconv. The Unoconv repository hasn't been updated since
then and the most recent release was on 2019-10-21, so archiving the
repository presumably means that the project is now unsupported. This
deprecates the formula accordingly.