percona-server: update livecheck
The existing `livecheck` block for `percona-server` fetches the HTML
for version options but it's currently failing because upstream has
changed the `value` attribute of the `option` elements to use a
format like `1.2.3|percona-server`, which the existing regex does not
match. This updates the regex to also allow `|` as an ending
delimiter, which resolves the issue.