mono-libgdiplus: update livecheck
The existing `livecheck` block for `mono-libgdiplus` presumably checks
the directory listing page where the `stable` tarball is found but
that URL is no longer accessible, so the check returns an `Unable to
get versions` error. The `stable` tarball is still accessible, it's
just that we can't fetch a listing of the files anymore.

This updates the `livecheck` block to check WineHQ GitLab releases,
which link to the related tarball from dl.winehq.org.