libatomic_ops: update urls
The upstream GitHub repository for `libatomic_ops` no longer contains
any releases, so the existing `stable` URL now returns a 404 (Not
Found) response. The repository was transferred to the bdwgc
organization and upstream updated the `README` in the existing
repository to link to the new repository, so this updates the URLs
accordingly.