otf2bdf 3.1_p1
I haven't been able to identify a way of finding version information
for `otf2bdf` from the current `stable` source, so I left it alone
for some time. We recently discussed checking the GitHub repository's
tags but this doesn't align with the `stable` source.

The software hasn't been meaningfully updated since 2015 but the
GitHub repository contains a few commits from 2021. The `3.1` tag is
for the code as it was in 2015 but there's also a `3.1_p1` tag that
includes the 2021 changes (the tags were created out of order, so
it looks like `3.1` is newer than `3.1_p1`).

This updates the formula to `3.1_p1`, using the tag tarball from
GitHub, and adds a `livecheck` block that checks the Git tags. The
build was failing on Linux CI with an `ft2build.h` error, due to
`otf2bdf.c` using `#include <ft2build.h>` instead of
`<freetype2/ft2build.h>`, but the `CFLAGS` snippet from other formulae
appears to resolve the error here as well.

Co-authored-by: Sean Molenaar <SMillerDev@users.noreply.github.com>