prips 1.2.2
This updates `prips` to the latest version, 1.2.2.

The existing `livecheck` block is returning an `Unable to get
versions` error, as the homepage no longer contains the expected
text. However, the upstream website now provides a download page that
links to the `stable` tarball, so this updates the `livecheck` block
to check that instead.