jjui v0.8.8 (new formula)
Add jjui, a TUI for jujutsu VCS.
See https://github.com/idursun/jjui and https://jj-vcs.github.io/jj/latest.

Being a TUI, I am testing against the output of jjui --version for the lack of a
better test, as well as the failure of the program when run in a directory that
is not jj repository.

It seemed reasonnable to depend on jj for a functional installation.