shadowenv: add livecheck
By default, livecheck checks the Git tags for `shadowenv` and this is
returning 3.1.0 as newest but there isn't a related release almost a
month after tagging. While we don't use release assets in the formula,
upstream directs users to either install through Homebrew or download
a binary from GitHub releases.

This adds a `livecheck` block that uses the `GithubLatest` strategy,
so we only update to versions with a release.