s6-rc 0.5.6.0 (new formula)
https://skarnet.org/software/s6-rc/index.html: s6-rc is a service
manager for s6-based systems, i.e. a suite of programs that can
start and stop services, both long-running daemons and one-time
initialization scripts, in the proper order according to a dependency
tree.

This formula depends on the existing "s6" formula and can be used to
control s6 supervision trees.