tiledb: add livecheck
By default livecheck checks the Git tags for `tiledb` but this is
now incorrectly returning 10040-azr-osx-lnxto-ga as the latest
version, from a `dlh/ch10040-azr-osx-lnxto-ga` tag. This adds a
`livecheck` block that uses the standard regex for Git tags like
`1.2.3`/`v1.2.3`, which will match stable tags in this repository
and omit unstable tags (e.g., `1.2.3-rc0`) as well as one-off oddities
like the aforementioned tag.