onigmo: add livecheck
By default, livecheck checks the Git tags for `onigmo` but it is
incorrectly returning 6.2.0_for_bregonig as the latest version, as
upstream creates `_for_bregonig` tags for some versions. The formula
uses a GitHub release asset in the `stable` URL, so this adds a
`livecheck` block that uses the `GithubLatest` strategy (i.e., the
release tags use an `Onigmo-1.2.3` format).