liquibase: update livecheck
The existing `livecheck` block for `liquibase` is returning an
`Unable to get versions` error, as the URL now redirects to
https://www.liquibase.com/download-pro and that page links to a Pro
tarball instead of the OSS distribution that the formula uses. The OSS
download page can be found at https://www.liquibase.com/download-oss
but the tarball link is hosted on package.liquibase.com, not GitHub.

This resolves the issue by updating the `livecheck` block to simply
use the `GithubLatest` strategy, as the formula is using a GitHub
release asset.